refactor: like 패키지 경로 변경
