+++ src/main/java/com/mobble/mobbleserver/domain/like/articleLike/dto/response/ArticleLikeCountResponseDto.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.like.articleLike.dto.response;
+package com.mobble.mobbleserver.domain.like.articleLike.dto.response;
 
 public record ArticleLikeCountResponseDto(Long articleId, int likeCount) {
     public static ArticleLikeCountResponseDto toDto(Long articleId, int count) {
