+++ src/main/java/com/mobble/mobbleserver/domain/like/articleLike/dto/response/ArticleLikeToggleResponseDto.java
@@ -1,8 +1,8 @@
-package com.mobble.mobbleserver.like.articleLike.dto.response;
+package com.mobble.mobbleserver.domain.like.articleLike.dto.response;
 
 import com.mobble.mobbleserver.domain.article.entity.Article;
 import com.mobble.mobbleserver.domain.member.entity.Member;
-import com.mobble.mobbleserver.like.LikeType;
+import com.mobble.mobbleserver.domain.like.LikeType;
 
 public record ArticleLikeToggleResponseDto(
         LikeType targetType,
