+++ src/main/java/com/mobble/mobbleserver/domain/like/clubLike/entity/ClubLike.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.like.clubLike.entity;
+package com.mobble.mobbleserver.domain.like.clubLike.entity;
 
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import jakarta.persistence.*;
