+++ src/main/java/com/mobble/mobbleserver/domain/like/commentLike/repository/CommentLikeRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.like.commentLike.repository;
+package com.mobble.mobbleserver.domain.like.commentLike.repository;
 
-import com.mobble.mobbleserver.like.commentLike.entity.CommentLike;
+import com.mobble.mobbleserver.domain.like.commentLike.entity.CommentLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentLikeRepository extends JpaRepository<CommentLike, Long> {
